fix build error

+++ src/compiler.c
@@ -107,7 +107,7 @@ static void emitReturn() {
 
 static uint8_t makeConstant(Value value) {
   int constant = addConstant(currentChunk(), value);
-  if (constant > UINT8_MAX) {
+  if (constant > (int)UINT8_MAX) {
     error("Too many constants in one chunk.");
     return 0;
   }
